# labels/cidr: use netip types to improve GetCIDRLabels and IPStringToLabel performace